refactor(BA-6630): move AgentInfo and handler session data to the data layer
Cut the remaining repositories/scheduler -> sokovan back-edges by relocating
AgentInfo and HandlerSessionData/HandlerKernelData into data/sokovan, and move
the AgentInfo.from_meta_and_occupancy factory into its only caller,
SessionProvisioner._build_agent_info.
