+++ tests/unit/manager/sokovan/scheduler/provisioner/selectors/test_failed_agent_filtering.py
@@ -8,6 +8,7 @@ from decimal import Decimal
 import pytest
 
 from ai.backend.common.types import AgentId, ClusterMode, ResourceSlot, SessionId, SessionTypes
+from ai.backend.manager.data.sokovan import AgentInfo
 from ai.backend.manager.sokovan.scheduler.provisioner.selectors.concentrated import (
     ConcentratedAgentSelector,
 )
@@ -15,7 +16,6 @@ from ai.backend.manager.sokovan.scheduler.provisioner.selectors.dispersed import
     DispersedAgentSelector,
 )
 from ai.backend.manager.sokovan.scheduler.provisioner.selectors.selector import (
-    AgentInfo,
     AgentSelectionConfig,
     AgentSelectionCriteria,
     AgentSelector,
