+++ tests/unit/manager/sokovan/scheduler/provisioner/selectors/test_selector_edge_cases.py
@@ -15,6 +15,7 @@ from ai.backend.common.types import (
     SessionId,
     SessionTypes,
 )
+from ai.backend.manager.data.sokovan import AgentInfo
 from ai.backend.manager.sokovan.scheduler.provisioner.selectors.concentrated import (
     ConcentratedAgentSelector,
 )
@@ -26,7 +27,6 @@ from ai.backend.manager.sokovan.scheduler.provisioner.selectors.roundrobin impor
     RoundRobinAgentSelector,
 )
 from ai.backend.manager.sokovan.scheduler.provisioner.selectors.selector import (
-    AgentInfo,
     AgentSelectionConfig,
     AgentSelectionCriteria,
     AgentStateTracker,
