+++ tests/unit/manager/sokovan/scheduler/test_scheduler.py
@@ -16,6 +16,7 @@ from ai.backend.common.types import (
     SessionTypes,
 )
 from ai.backend.manager.data.sokovan import (
+    AgentInfo,
     KernelWorkload,
     SchedulingConfig,
     SessionWorkload,
@@ -29,7 +30,6 @@ from ai.backend.manager.sokovan.scheduler.provisioner.selectors.exceptions impor
     NoCompatibleAgentError,
 )
 from ai.backend.manager.sokovan.scheduler.provisioner.selectors.selector import (
-    AgentInfo,
     AgentSelection,
     AgentSelectionConfig,
     AgentSelectionCriteria,
